processing/pyproject.toml
[project]
name = "bashkir-rag"
version = "0.1.0"
description = "RAG по документам Башкирэнерго | RTX 3060"
requires-python = ">=3.11,<3.12"
dependencies = [
    # Основные ML/DL библиотеки
    "torch==2.5.1",
    "torchvision==0.20.1",
    "torchaudio==2.5.1",

    # Langchain и связанные библиотеки
    "langchain==0.3.20",
    "langchain-community==0.3.19",
    "langchain-core==0.3.41",
    "langchain-ollama==0.2.2",
    "langchain-huggingface==0.1.0",
    "langchain-experimental",
    # Эмбеддинги
    "sentence-transformers==3.1.1",
    "qdrant-client==1.12.0",
    
    # OCR и обработка PDF
    "pymupdf==1.24.10",
    "pdf2image==1.17.0",
    "pytesseract==0.3.13",
    "unstructured[all-docs]==0.15.12",
    "pillow==10.4.0",
    
    
    # Вспомогательные библиотеки
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "black>=24.10.0",
    "flake8>=7.1.1",
]